BIG UPDATE!
rnmp3 has undergone a MAJOR update!  Now, in addition to being able to run it by passing the path containing the files to rename it can be run without passing a path and it will pop up a KDE Plasma path selection dialog for selecting the path to the files for renaming.  Speaking of renaming, a LOT of work has gone into making the renaming algorithm to make it more flexible.  Now it can rename successfully to the format "Artist - Album - Track Number - Title.mp3" if the filename is "tracknumber title.mp3", "track number-title.mp3" and a few permutations of "Artist(Album)-tracknumber-title.mp3"!  The script is not perfect as far as managing the spaces, but if you had any idea of what it took to get it to this point....  Once it does it's thing, it'll pop up a dialog summarizing the work done and asking if you want to do more, and acting accordingly, depending upon your choice.  Yes brings up the path selection dialog for the next file(s) to rename whereas No simply quits.