Konversation-iroffer player script!
This is a modification of the Konversation /media (/audio & /video too) script that goes in the /usr/share/konversation/scripts/ directory.  This one has been modified to not only detect the currently playing filer and player, but will also find the file's pack number in your iroffer file list and advertise it for download!  MAKE SURE to edit the file and path names in the script in order for it to work with your setup!