UPDATED script!
Ok, so I have been working with ChatGPT for over 4 days straight trying to tweak this script out to work properly, and it is now working better than it ever has before.  It still requires the KDE Plasma DE, and can be run from the terminal or dolphin and will rename all the files in the directory (and all subdirectories) to the spec "Artist - Album - Track Number - Track Title.mp3" or "Artist - Track Title.mp3".  It only works on mp3 files currently, and is still in the early testing phase.  It requires the following directory structure:  ".../mp3/Artist/Album/..." where Artist = the Artist's name and Album = the name of the CD/Album.  It relies on this directory structure to know which way to rename the files.